+++ Monster.cpp
@@ -12,7 +12,7 @@ bool Monster::isAlive() const
     return health > 0;
 }
 
-Monster::Monster(const std::string name, const int health, const int damage, const double atkcooldown, const double defense) : name(name), health(health), damage(damage), atkcooldown(atkcooldown), defense(defense)
+Monster::Monster(const std::string name, const int health, const int damage, const double atkcooldown, const int defense) : name(name), health(health), damage(damage), atkcooldown(atkcooldown), defense(defense)
 {
 }
 
@@ -36,7 +36,7 @@ double Monster::getAttackCoolDown() const
     return atkcooldown;
 }
 
-double Monster::getDefense() const
+int Monster::getDefense() const
 {
     return defense;
 }
@@ -91,5 +91,5 @@ void Monster::SufferDamage(int damageRecieved)
 Monster Monster::parse(const std::string &fileName)
 {
     JSON data = JSON::parseFromFile(fileName);
-    return Monster(data.get<std::string>("name"), data.get<int>("health_points"), data.get<int>("damage"), data.get<double>("attack_cooldown"), data.get<double>("defense"));
+    return Monster(data.get<std::string>("name"), data.get<int>("health_points"), data.get<int>("damage"), data.get<double>("attack_cooldown"), data.get<int>("defense"));
 }
